Fix createRoute parameter type in SearchDetailScreen

+++ app/src/main/java/com/cabovianco/reelio/presentation/navigation/Screen.kt
@@ -4,7 +4,7 @@ sealed class Screen(val route: String) {
     data object SearchScreen : Screen("search")
 
     data object SearchDetailScreen : Screen("search/{id}") {
-        fun createRoute(id: Long) = "search/$id"
+        fun createRoute(id: Int) = "search/$id"
     }
 
     data object BookmarkScreen : Screen("bookmark")
